move stuff from test/utils to ts/
