+++ test/Deposit.spec.ts
@@ -1,4 +1,3 @@
 describe("DepositManager", async function() {
-
     it("should allow depositing 2 leaves in a subtree and merging it", async () => {});
 });
