+++ test/rollup-utils.spec.ts
@@ -1,16 +1,8 @@
 import { assert } from "chai";
-import {
-    CreateAccount,
-    Account,
-    Transaction,
-    DropTx,
-    Usage,
-    BurnConsentTx,
-    BurnExecutionTx
-} from "../../scripts/helpers/interfaces";
-import { DummyECDSASignature } from "../../scripts/helpers/constants";
-import { RollupUtilsFactory } from "../../types/ethers-contracts/RollupUtilsFactory";
-import { RollupUtils } from "../../types/ethers-contracts/RollupUtils";
+import { TxTransfer, TxCreate, TxBurnConsent, TxBurnExecution } from "../ts/tx";
+import { EMPTY_ACCOUNT } from "../ts/state_account";
+import { RollupUtilsFactory } from "../types/ethers-contracts/RollupUtilsFactory";
+import { RollupUtils } from "../types/ethers-contracts/RollupUtils";
 import { ethers } from "@nomiclabs/buidler";
 
 describe("RollupUtils", async function() {
@@ -21,14 +13,7 @@ describe("RollupUtils", async function() {
     });
 
     it("test account encoding and decoding", async function() {
-        const account: Account = {
-            ID: 1,
-            tokenType: 2,
-            balance: 3,
-            nonce: 4,
-            burn: 0,
-            lastBurn: 0
-        };
+        const account = EMPTY_ACCOUNT;
 
         const accountBytes = await RollupUtilsInstance.BytesFromAccountDeconstructed(
             account.ID,
@@ -48,17 +33,7 @@ describe("RollupUtils", async function() {
         assert.equal(regeneratedAccount["4"].toNumber(), account.burn);
         assert.equal(regeneratedAccount["5"].toNumber(), account.lastBurn);
 
-        const tx: Transaction = {
-            txType: Usage.Transfer,
-            fromIndex: 1,
-            toIndex: 2,
-            tokenType: 1,
-            amount: 10,
-            fee: 1,
-            signature:
-                "0x1ad4773ace8ee65b8f1d94a3ca7adba51ee2ca0bdb550907715b3b65f1e3ad9f69e610383dc9ceb8a50c882da4b1b98b96500bdf308c1bdce2187cb23b7d736f1b",
-            nonce: 0
-        };
+        const tx = TxTransfer.rand().extended();
 
         const txBytes = await RollupUtilsInstance.BytesFromTxDeconstructed(
             tx.txType,
@@ -79,12 +54,7 @@ describe("RollupUtils", async function() {
         assert.equal(txData.amount.toString(), tx.amount.toString());
     });
     it("test createAccount utils", async function() {
-        const tx: CreateAccount = {
-            txType: Usage.CreateAccount,
-            accountID: 1,
-            stateID: 1,
-            tokenType: 1
-        };
+        const tx = TxCreate.rand().extended();
         const txBytes = await RollupUtilsInstance.BytesFromCreateAccountNoStruct(
             tx.txType,
             tx.accountID,
@@ -104,15 +74,7 @@ describe("RollupUtils", async function() {
         await RollupUtilsInstance.DecompressManyCreateAccount(txs);
     });
     it("test airdrop utils", async function() {
-        const tx: DropTx = {
-            txType: Usage.Airdrop,
-            fromIndex: 1,
-            toIndex: 1,
-            tokenType: 1,
-            nonce: 2,
-            amount: 10,
-            signature: DummyECDSASignature
-        };
+        const tx = TxTransfer.rand().extended();
         const signBytes = await RollupUtilsInstance.AirdropSignBytes(
             tx.txType,
             tx.fromIndex,
@@ -130,10 +92,7 @@ describe("RollupUtils", async function() {
         );
         const result = await RollupUtilsInstance.AirdropFromBytes(txBytes);
         assert.equal(Number(result.toIndex), tx.toIndex);
-        await RollupUtilsInstance.CompressAirdropFromEncoded(
-            txBytes,
-            tx.signature
-        );
+        await RollupUtilsInstance.CompressAirdropFromEncoded(txBytes, "0x00");
         const txs = await RollupUtilsInstance.CompressManyAirdropFromEncoded([
             txBytes,
             txBytes
@@ -141,16 +100,7 @@ describe("RollupUtils", async function() {
         await RollupUtilsInstance.DecompressManyAirdrop(txs);
     });
     it("test transfer utils", async function() {
-        const tx: Transaction = {
-            txType: Usage.Transfer,
-            fromIndex: 1,
-            toIndex: 2,
-            tokenType: 1,
-            nonce: 3,
-            amount: 1,
-            fee: 1,
-            signature: DummyECDSASignature
-        };
+        const tx = TxTransfer.rand().extended();
         const signBytes = await RollupUtilsInstance.getTxSignBytes(
             tx.txType,
             tx.fromIndex,
@@ -168,24 +118,15 @@ describe("RollupUtils", async function() {
             tx.amount,
             tx.fee
         );
-        await RollupUtilsInstance.CompressTransferFromEncoded(
-            txBytes,
-            tx.signature
-        );
+        await RollupUtilsInstance.CompressTransferFromEncoded(txBytes, "0x00");
         const txs = await RollupUtilsInstance.CompressManyTransferFromEncoded(
             [txBytes, txBytes],
-            [tx.signature, tx.signature]
+            ["0x00", "0x00"]
         );
         await RollupUtilsInstance.DecompressManyTransfer(txs);
     });
     it("test burn consent utils", async function() {
-        const tx: BurnConsentTx = {
-            txType: Usage.BurnConsent,
-            fromIndex: 1,
-            amount: 5,
-            nonce: 3,
-            signature: DummyECDSASignature
-        };
+        const tx = TxBurnConsent.rand().extended();
         const signBytes = await RollupUtilsInstance.BurnConsentSignBytes(
             tx.txType,
             tx.fromIndex,
@@ -206,10 +147,7 @@ describe("RollupUtils", async function() {
         await RollupUtilsInstance.DecompressManyBurnConsent(txs);
     });
     it("test burn execution utils", async function() {
-        const tx: BurnExecutionTx = {
-            txType: Usage.BurnExecution,
-            fromIndex: 5
-        };
+        const tx = TxBurnExecution.rand().extended();
         const txBytes = await RollupUtilsInstance.BytesFromBurnExecutionNoStruct(
             tx.txType,
             tx.fromIndex
