+++ ts/tree/hasher.ts
